requirements.txt
discord.py
firebase_admin
python-decouple
discord.py[voice]
youtube_dl